+++ src/ts/Test/Unit/Managers/CampaignManagerTest.ts
@@ -32,6 +32,7 @@ import { ProgrammaticVastParser } from 'Parsers/ProgrammaticVastParser';
 import { AdMobSignalFactory } from 'AdMob/AdMobSignalFactory';
 import { AdMobSignal } from 'Models/AdMobSignal';
 import { CacheBookkeeping } from 'Utilities/CacheBookkeeping';
+import { AndroidDeviceInfo } from 'Models/AndroidDeviceInfo';
 
 import ConfigurationAuctionPlc from 'json/ConfigurationAuctionPlc.json';
 import OnProgrammaticMraidPlcCampaignEmpty from 'json/OnProgrammaticMraidPlcCampaignEmpty.json';
@@ -194,7 +195,7 @@ describe('CampaignManager', () => {
         focusManager = new FocusManager(nativeBridge);
         wakeUpManager = new WakeUpManager(nativeBridge, focusManager);
         request = new Request(nativeBridge, wakeUpManager);
-        deviceInfo = new DeviceInfo(nativeBridge);
+        deviceInfo = new AndroidDeviceInfo(nativeBridge);
         metaDataManager = new MetaDataManager(nativeBridge);
         thirdPartyEventManager = new ThirdPartyEventManager(nativeBridge, request);
         sessionManager = new SessionManager(nativeBridge);
@@ -1042,7 +1043,7 @@ describe('CampaignManager', () => {
                     assert.equal(triggeredCampaign.getCorrelationId(), '0zGg2TfRsBNbqlc7AVdhLAw');
                     assert.deepEqual((<MRAIDCampaign>triggeredCampaign).getResourceUrl(), new HTML('https://img.serveroute.com/mini_8ball_fast/inlined.html', triggeredCampaign.getSession()));
                     assert.deepEqual((<MRAIDCampaign>triggeredCampaign).getDynamicMarkup(), 'var markup = \'dynamic\';');
-                    assert.deepEqual((<MRAIDCampaign>triggeredCampaign).getTrackingEventUrls(), {
+                    assert.deepEqual((<MRAIDCampaign>triggeredCampaign).getTrackingUrls(), {
                         "impression": [
                             "http://test.impression.com/blah1",
                             "http://test.impression.com/blah2",
@@ -1078,7 +1079,7 @@ describe('CampaignManager', () => {
                     assert.equal(triggeredCampaign.getSeatId(), 902);
                     assert.equal(triggeredCampaign.getCorrelationId(), 'zGg2TfRsBNbqlc7AVdhLAw');
                     assert.deepEqual((<MRAIDCampaign>triggeredCampaign).getResource(), '<div>markup</div>');
-                    assert.deepEqual((<MRAIDCampaign>triggeredCampaign).getTrackingEventUrls(), {impression: ['https://ads-brand-postback.unityads.unity3d.com/brands/2000/%ZONE%/impression/common?data=Kz2J']});
+                    assert.deepEqual((<MRAIDCampaign>triggeredCampaign).getTrackingUrls(), {impression: ['https://ads-brand-postback.unityads.unity3d.com/brands/2000/%ZONE%/impression/common?data=Kz2J']});
                 });
             });