+++ src/ts/Utilities/Cache.ts
@@ -361,9 +361,7 @@ export class Cache {
             } else if(Request.RedirectResponseCodes.exec(responseCode.toString())) {
                 this.sendDiagnostic(CacheDiagnosticEvent.REDIRECTED, callback);
                 this._cacheBookkeeping.removeFileEntry(callback.fileId);
-                if(size > 0) {
-                    this._nativeBridge.Cache.deleteFile(callback.fileId);
-                }
+                this._nativeBridge.Cache.deleteFile(callback.fileId);
                 const location = Request.getHeader(headers, 'location');
                 if(location) {
                     let fileId = callback.fileId;