+++ src/ts/Utilities/IosUtils.ts
@@ -1,7 +1,13 @@
 export class IosUtils {
 
-    public static isAppSheetBroken(osVersion: string): boolean {
-        return osVersion.match(/^8\.[0-3]/) !== null;
+    public static isAppSheetBroken(osVersion: string, model: string): boolean {
+        if(!model.match(/ipad/i) && osVersion.match(/^11\.[0-3]/)) {
+            return true;
+        } else if(osVersion.match(/^8\.[0-3]/)) {
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public static hasVideoStallingApi(osVersion: string): boolean {