+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+# Thu Mar  1 10:39:02 UTC 2018
+
+* Validate and encode all incoming asset URLs
+* Fix cache redirections on 2.2.0
+* Show time to close / skip for MRAID overlays on group 18 and 19
+* Fix creative slot colliding with video slot for VPAID
+* Fixed issue with how setViewFrame was being used
+
+# Wed Feb 28 13:55:33 UTC 2018
+
+* Fixes for openUrl and sendEvent for VPAID
+* New AdMob spam signals for SDK 2.2.0
+* Refactor and add missing device infos
+* Overlay progress bar AB test for groups 16 & 17
+* Remove lunar theme AB test
+
+# Fri Feb 23 15:10:28 UTC 2018
+
+* Fix MOAT init & closing due to bad merge
+
+# Fri Feb 23 10:27:13 UTC 2018
+
+* Purchasing utilities cleanup
+* Disable app sheet on iPhone iOS 11
+* Fix orientation issues for Promo Android
+* Display ads: fixed ABT-373 and ABT-374 Cleanups, simplification of adunit.
+* Fix the view rendering and add system interrupt observer for iOS foreground and background events to VPAIDAdUnit
+* VPAID container fixes
+* Fix hybrid tests & disable 2 broken PurchasingUtilities tests
+* Add iOS screenScale to ad request query parameters
+* Fix missing MOAT stylesheet 
+
+# Thu Feb 22 08:57:26 UTC 2018
+
+* Preparations for new double webview APIs in 2.2.0 release
+* Preparations for new IAP promo APIs in 2.2.0 release
+* Removal of VPAID and display ad support from all pre-2.2.0 releases
+* Remove video overlay test from groups 8 and 10
+* Start random download button color test on groups 8 and 10
+
 # Fri Feb 16 10:18:25 UTC 2018
 
 * Add ad unit style object to operative events